Fixes resid corr cofe and completes refactor of the demixingresults so that we can import a and c from any pipeline and view the quality of demixing with respect to pmd
