+++ masknmf/demixing/__init__.py
@@ -1,6 +1,5 @@
-from .signal_demixer import SignalDemixer, InitializingState, DemixingState
+from .signal_demixer import SignalDemixer, InitializingState, DemixingState, DemixingResults
 from .demixing_arrays import (
-    DemixingResults,
     ACArray,
     ColorfulACArray,
     FluctuatingBackgroundArray,
@@ -8,5 +7,5 @@ from .demixing_arrays import (
     ResidCorrMode,
 )
 from .background_estimation import RingModel
-from .demixing_utils import torch_sparse_to_scipy_coo
+from .demixing_utils import torch_sparse_to_scipy_coo, ndarray_to_torch_sparse_coo
 from .filters import high_pass_filter_batch
