+++ masknmf/demixing/demixing_arrays.py
@@ -1,6 +1,8 @@
 from typing import *
 import numpy as np
 from enum import Enum
+
+from masknmf import display
 from masknmf.arrays.array_interfaces import FactorizedVideo
 from masknmf.compression import PMDArray
 import torch
@@ -1254,205 +1256,3 @@ class StandardCorrelationImages(FactorizedVideo):
         product = product.cpu().numpy().astype(self.dtype).squeeze()
         return product
 
-
-class DemixingResults:
-    def __init__(
-        self,
-        u_sparse: torch.sparse_coo_tensor,
-        q: torch.tensor,
-        v: torch.tensor,
-        a: torch.sparse_coo_tensor,
-        c: torch.tensor,
-        b: torch.tensor,
-        residual_correlation_image: ResidualCorrelationImages,
-        standard_correlation_image: StandardCorrelationImages,
-        background_to_signal_correlation_image: StandardCorrelationImages,
-        order: str,
-        data_shape: tuple[int, int, int],
-        device="cpu",
-    ):
-        """
-        This class provides a convenient way to export all demixing result as array-like objects.
-        Args:
-            u_sparse (torch.sparse_coo_tensor): shape (pixels, rank 1)
-            q (torch.tensor): shape (rank 2, rank 2)
-            v (torch.tensor): shape (rank 2, num_frames)
-            a (torch.sparse_coo_tensor): shape (pixels, number of neural signals)
-            c (torch.tensor): shape (number of frames, number of neural signals)
-            b (torch.tensor): shape (pixels)
-            residual_correlation_image (ResidualCorrelationImages): Shape (number of neural signals, FOV dim 1, FOV dim 2)
-                The residual corr img corresponding to these demixing results
-            standard_correlation_image (StandardCorrelationImages): Shape (number of neural signals,
-                FOV dim 1, FOV dim 2): The standard corr img corresponding to these demixing results
-            order (str): order used to reshape data from 2D to 1D
-            data_shape (tuple): (number of frames, field of view dimension 1, field of view dimension 2)
-            device (str): 'cpu' or 'cuda'. used to manage where the tensors reside
-        """
-        self._device = device
-        self._order = order
-        self._shape = data_shape
-        self._u_sparse = u_sparse.to(device)
-        self._q = q
-        self._v = v
-        self._a = a
-        self._c = c
-
-        ## Store the critical values for each of these
-        self._std_corr_img_mean = standard_correlation_image.movie_mean
-        self._std_corr_img_normalizer = standard_correlation_image.movie_normalizer
-
-        self._bkgd_std_corr_img_mean = background_to_signal_correlation_image.movie_mean
-        self._bkgd_std_corr_img_normalizer = background_to_signal_correlation_image.movie_normalizer
-
-        self._resid_corr_img_support_values = residual_correlation_image.support_correlation_values
-        self._resid_corr_img_mean = residual_correlation_image.residual_movie_mean
-        self._resid_corr_img_normalizer = residual_correlation_image.residual_movie_normalizer
-
-        if self.order == "C":
-            self._baseline = b.reshape((self.shape[1], self.shape[2]))
-        elif self.order == "F":
-            # Note we swap 1 and 2 here
-            self._baseline = b.reshape((self.shape[2], self.shape[1])).T
-
-        #Move all tracked tensors to desired location so everything is on one device
-        self.to(device)
-
-    @property
-    def standard_correlation_image(self) -> StandardCorrelationImages:
-        return StandardCorrelationImages(self._u_sparse,
-                                         self._v,
-                                         self._c,
-                                         self._std_corr_img_mean,
-                                         self._std_corr_img_normalizer,
-                                         (self._shape[1], self._shape[2]),
-                                         order=self.order)
-
-    @property
-    def background_to_signal_correlation_image(self) -> StandardCorrelationImages:
-        return StandardCorrelationImages(self._u_sparse,
-                                         self._q[0] @ self._q[1],
-                                         self._c,
-                                         self._bkgd_std_corr_img_mean,
-                                         self._bkgd_std_corr_img_normalizer,
-                                         (self._shape[1], self._shape[2]),
-                                         order=self.order)
-
-    @property
-    def residual_correlation_image(self) -> ResidualCorrelationImages:
-        return ResidualCorrelationImages(self._u_sparse,
-                                         self._v,
-                                         self._q,
-                                         self._a,
-                                         self._c,
-                                         self._resid_corr_img_support_values,
-                                         self._resid_corr_img_mean,
-                                         self._resid_corr_img_normalizer,
-                                         (self._shape[1], self._shape[2]),
-                                         mode=ResidCorrMode.RESIDUAL,
-                                         order=self._order)
-    @property
-    def shape(self):
-        return self._shape
-
-    @property
-    def order(self):
-        return self._order
-
-    @property
-    def device(self):
-        return self._device
-
-    def to(self, new_device):
-        self._device = new_device
-        self._u_sparse = self._u_sparse.to(self.device)
-        self._q = (self._q[0].to(self.device), self._q[1].to(self.device))
-        self._v = self._v.to(self.device)
-        self._a = self._a.to(self.device)
-        self._c = self._c.to(self.device)
-        self._baseline = self.baseline.to(self.device)
-        self._std_corr_img_mean = self._std_corr_img_mean.to(self.device)
-        self._std_corr_img_normalizer = self._std_corr_img_normalizer.to(self.device)
-
-        self._bkgd_std_corr_img_mean = self._bkgd_std_corr_img_mean.to(self.device)
-        self._bkgd_std_corr_img_normalizer = self._bkgd_std_corr_img_normalizer.to(self.device)
-
-        self._resid_corr_img_support_values = self._resid_corr_img_support_values.to(self.device)
-        self._resid_corr_img_mean = self._resid_corr_img_mean.to(self.device)
-        self._resid_corr_img_normalizer = self._resid_corr_img_normalizer.to(self.device)
-
-    @property
-    def fov_shape(self) -> Tuple[int, int]:
-        return self.shape[1:3]
-
-    @property
-    def num_frames(self) -> int:
-        return self.shape[0]
-
-    @property
-    def u(self) -> torch.sparse_coo_tensor:
-        return self._u_sparse
-
-    @property
-    def baseline(self) -> torch.tensor:
-        return self._baseline
-
-    @property
-    def q(self) -> torch.tensor:
-        return self._q
-
-    @property
-    def v(self) -> torch.tensor:
-        return self._v
-
-    @property
-    def a(self) -> torch.sparse_coo_tensor:
-        return self._a
-
-    @property
-    def c(self) -> torch.tensor:
-        return self._c
-
-    @property
-    def ac_array(self) -> ACArray:
-        """
-        Returns an ACArray using the tensors stored in this object
-        """
-        return ACArray(self.fov_shape, self.order, self.a, self.c)
-
-    @property
-    def pmd_array(self) -> PMDArray:
-        """
-        Returns a PMDArray using the tensors stored in this object
-        """
-        mean_img = torch.zeros(self.shape[1], self.shape[2], device=self.device)
-        var_img = torch.ones(self.shape[1], self.shape[2], device=self.device)
-        return PMDArray(
-            self.shape,
-            self.u,
-            self.v,
-            mean_img,
-            var_img,
-            device=self.device,
-            rescale=True,
-        )
-
-    @property
-    def fluctuating_background_array(self) -> FluctuatingBackgroundArray:
-        """
-        TODO: When the refactor is complete, need to return once again a Fluctuating Background Array
-        Returns a PMDArray using the tensors stored in this object
-        """
-        return FluctuatingBackgroundArray(self.fov_shape, self.order, self.u, self.q[0], self.q[1])
-
-    @property
-    def residual_array(self) -> ResidualArray:
-        return ResidualArray(
-            self.pmd_array,
-            self.ac_array,
-            self.fluctuating_background_array,
-            self.baseline,
-        )
-
-    @property
-    def colorful_ac_array(self) -> ColorfulACArray:
-        return ColorfulACArray(self.fov_shape, self.order, self.a, self.c)
